% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_and_pred.R
\name{both_cases_check}
\alias{both_cases_check}
\title{computes required quantiles and feeds them to case one  adn two functions
given by Keef (2013)}
\usage{
both_cases_check(alpha, beta, yi, ymnsi, q)
}
\arguments{
\item{alpha}{float, H&T alpha parameter}

\item{beta}{float, H&T beta parameter}

\item{yi}{vector of conditioning variable}

\item{ymnsi}{vector of dependent variable}

\item{q}{???? check Keef (2013) for this}
}
\value{
boolean, true if case on is satisfied
}
\description{
computes required quantiles and feeds them to case one  adn two functions
given by Keef (2013)
}
\keyword{internal}
