% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bivariate_dens.R
\name{emp_box_prob}
\alias{emp_box_prob}
\title{Counts proportion of points in box}
\usage{
emp_box_prob(point, data, x_div, y_div, x_mid_l, x_mid)
}
\description{
Takes the centre point of box and locations for division points, then calculates
the proportion of points in those boxes
}
\keyword{internal}
