% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/thresh_select.R
\name{energy.thrsh.test}
\alias{energy.thrsh.test}
\title{conducts test of independence based on that of Wan (2019) to validate H&T
exceedance quantiles}
\usage{
energy.thrsh.test(data, qt.range, n.boot, n.boot.par, n.rep)
}
\arguments{
\item{data}{matrix of 2 columns containing to fit HT to}

\item{qt.range}{range of exceedance threshold quantiles to evaluate}

\item{n.boot}{size of bootstraps to take at each rep}

\item{n.boot.par}{size of boostrapped parameter samples to calculate at each rep}

\item{n.rep}{number of p value calculation reps per quantiles}
}
\value{
returns nothing
}
\description{
conducts test of independence based on that of Wan (2019) to validate H&T
exceedance quantiles
}
