% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/thresh_select.R
\name{eq_split}
\alias{eq_split}
\title{splits a vector into n equal chunks (with overlap if necessary)}
\usage{
eq_split(X, n)
}
\description{
splits a vector into n equal chunks (with overlap if necessary)
}
\keyword{internal}
