% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/thresh_select.R
\name{get_bands}
\alias{get_bands}
\title{split Y into given number of bands by X values}
\usage{
get_bands(sample, n_bands)
}
\description{
split Y into given number of bands by X values
}
\keyword{internal}
