% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/thresh_select.R
\name{get_t}
\alias{get_t}
\title{get test statistic t from quantiles and all residuals}
\usage{
get_t(cdfs, all_z)
}
\description{
get test statistic t from quantiles and all residuals
}
\keyword{internal}
