% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/thresh_select.R
\name{get_X_and_Z}
\alias{get_X_and_Z}
\title{get sorted X-u and Z for a given sample & fit conditions}
\usage{
get_X_and_Z(sample, qu, theta0)
}
\description{
get sorted X-u and Z for a given sample & fit conditions
}
\keyword{internal}
