% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bivariate_dens.R
\name{grid.dens}
\alias{grid.dens}
\title{Gridded density estimate using H&T for extremes}
\usage{
grid.dens(data, q.marginal, q.cond, xlim, nx, ny)
}
\arguments{
\item{data}{matrix of data in original margins}

\item{q.marginal}{threshold exceedance quantile to use for margins}

\item{q.cond}{threshold exceedance quantile to use for conditional model}

\item{xlim}{x limits}

\item{nx}{number of division points for xgrid}

\item{ny}{number of division points for ygrid}
}
\value{
dataframe of densities and probabilities
}
\description{
Calculates a grid based density estimate for bivariate data, by using the
empirical density for regions in the body of the data and H&T (2004) for
extreme regions

GPD tails are fitted to marginal empirical distributions to transform to
laplace margins
}
