% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_and_pred.R
\name{ht.pred}
\alias{ht.pred}
\title{Prediction for conditional extremes}
\usage{
ht.pred(
  v,
  theta,
  Z,
  n_pred,
  Y = T,
  points = F,
  plot = T,
  inds = F,
  col = "darkred"
)
}
\arguments{
\item{v}{float, value of conditioned variable to predict above}

\item{theta}{vector, pars from H&T fit to use in prediction}

\item{Z}{vector, residuals from H&T fit to use in prediction}

\item{n_pred}{int, number of predictions to make}

\item{Y}{bool, true if we are modeling Y|X, false if X|Y}

\item{plot}{bool, set to true to see visualisation}

\item{col}{string, sets the color of plotted predictions}
}
\value{
matrix of predicted X,Y points
}
\description{
generates prediction points for a given H&T model, with options to plot and/or
output prediction points
}
