% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/thresh_select.R
\name{indv.thrsh.test}
\alias{indv.thrsh.test}
\title{T value indepedence test for individual threshold}
\usage{
indv.thrsh.test(
  sample,
  qu,
  theta0,
  n_bands = 5,
  sig = 0.05,
  plots = T,
  k_min = 5,
  nr = 1000
)
}
\arguments{
\item{sample}{matrix of two columns containing sample data}

\item{qu}{candidate threshold quantile for conditioning variable}

\item{theta0}{vector of initial H&T parameter values to be used in optimisation}

\item{n_bands}{integer number of bands to split data into}

\item{sig}{float level of sigficance to test against}

\item{plots}{bool, true to show diagnostic plots}

\item{k_min}{int min number of points acceptable in each band}

\item{nr}{int number of randomised bands}
}
\value{
list of resulting p-value boolean decision of accept/reject
}
\description{
uses difference metric idea to check independence of X-u and Z, thus checking
the suitability of qu as a threshold quantile
}
