% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{pspliced}
\alias{pspliced}
\title{Distribution function for GPD spliced with empirical}
\usage{
pspliced(x0, x, u, gpd_par)
}
\arguments{
\item{x0}{float to evaluate the distribution at}

\item{x}{vector of sample used to fit the distribution}

\item{u}{float, exceedance threshold}

\item{gpd_par}{vector of gpd scale and shape parameters}
}
\value{
float between \link{0,1}, estimated cdf at x0
}
\description{
spliced distribution which uses empirical cdf for below threshold and gpd
for above
}
