% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{q_spliced_dist_est}
\alias{q_spliced_dist_est}
\title{spliced inverse distribution which uses empirical cdf for below threshold
and gpd for above}
\usage{
q_spliced_dist_est(p, x, q, gpd_par)
}
\arguments{
\item{p}{float between \link{0,1} for which to find the corresponding quantile}

\item{x}{vector of sample used to fit the distribution}

\item{q}{float between \link{0,1}, exceedance quantile}

\item{gpd_par}{vector of gpd scale and shape parameters}
}
\value{
float, estimated pth quantile for data x with gpd tail
}
\description{
spliced inverse distribution which uses empirical cdf for below threshold
and gpd for above
}
