% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/thresh_select.R
\name{tval.thrsh.test}
\alias{tval.thrsh.test}
\title{uses difference metric idea to check independence of X-u and Z, thus checking
the suitability of qu as a threshold quantile
does this for a range of quantiles, using an adaptive number of bands at each
stage}
\usage{
tval.thrsh.test(sample, qrange, theta0, sig = 0.05, k_min = 10)
}
\arguments{
\item{sample}{matrix of two columns containing sample data}

\item{qrange}{candidate threshold quantiles for conditioning variable}

\item{theta0}{vector of initial H&T parameter values to be used in optimisation}

\item{sig}{float level of sigficance to test against}

\item{k_min}{int min number of points acceptable in each band}
}
\value{
no value is returned
}
\description{
uses difference metric idea to check independence of X-u and Z, thus checking
the suitability of qu as a threshold quantile
does this for a range of quantiles, using an adaptive number of bands at each
stage
}
